In the middle of reworking timers
Basically, the issue is that numerous actions call get_prev_deltatime(),
which is a method that causes a state change, and so all actions are not
getting the actual proper delta time since last reading. This should be
changed by introducing another supplementary timeclock and also have
unified single read action for an entire iteration, even if this costs a
little bit of accuracy.